+++ packages/core/src/redux/slices/__tests__/roomConnection.unit.ts
@@ -11,6 +11,7 @@ describe("roomConnectionSlice", () => {
                     signalEvents.roomJoined({
                         error: "room_locked",
                         selfId: "selfId",
+                        clientClaim: "clientClaim",
                         isLocked: true,
                     }),
                 );
@@ -26,8 +27,8 @@ describe("roomConnectionSlice", () => {
                 const result = roomConnectionSlice.reducer(
                     undefined,
                     signalEvents.roomJoined({
-                        error: "room_locked",
                         selfId: "selfId",
+                        clientClaim: "clientClaim",
                         isLocked: false,
                     }),
                 );
@@ -39,6 +40,25 @@ describe("roomConnectionSlice", () => {
                 });
             });
 
+            it("should set status to disconnected and populate the error if the there is an error", () => {
+                const result = roomConnectionSlice.reducer(
+                    undefined,
+                    signalEvents.roomJoined({
+                        error: "room_full",
+                        selfId: "selfId",
+                        isLocked: false,
+                    }),
+                );
+
+                expect(result).toEqual({
+                    status: "disconnected",
+                    session: null,
+                    error: "room_full",
+                });
+            });
+        });
+
+        describe("signalEvents.clientKicked", () => {
             it("should set status to kicked if the client is kicked", () => {
                 const result = roomConnectionSlice.reducer(
                     undefined,
@@ -54,6 +74,18 @@ describe("roomConnectionSlice", () => {
                 });
             });
         });
+
+        describe("signalEvents.disconnect", () => {
+            it("should set status to disconnected", () => {
+                const result = roomConnectionSlice.reducer(undefined, signalEvents.disconnect());
+
+                expect(result).toEqual({
+                    status: "disconnected",
+                    session: null,
+                    error: null,
+                });
+            });
+        });
     });
 
     describe("reactors", () => {
@@ -61,21 +93,23 @@ describe("roomConnectionSlice", () => {
             const x = () => oneOf(true, false);
 
             it.each`
-                organizationId | roomConnectionStatus | signalIdentified | localMediaStatus | expected
-                ${undefined}   | ${"initializing"}    | ${x()}           | ${"started"}     | ${false}
-                ${"orgId"}     | ${"initializing"}    | ${true}          | ${"started"}     | ${true}
-                ${"orgId"}     | ${"connected"}       | ${x()}           | ${"started"}     | ${false}
-                ${"orgId"}     | ${"initializing"}    | ${false}         | ${"starting"}    | ${false}
-                ${"orgId"}     | ${"initializing"}    | ${x()}           | ${"error"}       | ${false}
+                organizationId | roomConnectionStatus | signalIdentified | localMediaStatus | isNodeSdk | expected
+                ${undefined}   | ${"initializing"}    | ${x()}           | ${"started"}     | ${x()}    | ${false}
+                ${"orgId"}     | ${"initializing"}    | ${true}          | ${"started"}     | ${false}  | ${true}
+                ${"orgId"}     | ${"connected"}       | ${x()}           | ${"started"}     | ${x()}    | ${false}
+                ${"orgId"}     | ${"initializing"}    | ${false}         | ${"starting"}    | ${x()}    | ${false}
+                ${"orgId"}     | ${"initializing"}    | ${x()}           | ${"error"}       | ${false}  | ${false}
+                ${"orgId"}     | ${"initializing"}    | ${true}          | ${"error"}       | ${true}   | ${true}
             `(
-                "should return $expected when hasOrganizationIdFetched=$hasOrganizationIdFetched, roomConnectionStatus=$roomConnectionStatus, signalIdentified=$signalIdentified, localMediaStatus=$localMediaStatus",
-                ({ organizationId, roomConnectionStatus, signalIdentified, localMediaStatus, expected }) => {
+                "Should return $expected when organizationId=$organizationId, roomConnectionStatus=$roomConnectionStatus, signalIdentified=$signalIdentified, localMediaStatus=$localMediaStatus, isNodeSdk=$isNodeSdk",
+                ({ organizationId, roomConnectionStatus, signalIdentified, localMediaStatus, isNodeSdk, expected }) => {
                     expect(
                         selectShouldConnectRoom.resultFunc(
                             organizationId,
                             roomConnectionStatus,
                             signalIdentified,
                             localMediaStatus,
+                            isNodeSdk,
                         ),
                     ).toEqual(expected);
                 },